+++ src/Misc/Hooks.BugFixes.cpp
@@ -5,6 +5,7 @@
 #include <UnitClass.h>
 #include <ScenarioClass.h>
 #include <VoxelAnimClass.h>
+#include <BulletClass.h>
 
 #include <Utilities/Macro.h>
 #include <Utilities/Debug.h>
@@ -160,7 +161,7 @@ DEFINE_HOOK(0x4FB2DE, HouseClass_PlaceObject_HotkeyFix, 0x6)
 	GET(TechnoClass*, pObject, ESI);
 
 	pObject->ClearSidebarTabObject();
-	
+
 	return 0;
 }
 
@@ -173,7 +174,7 @@ DEFINE_HOOK(0x44377E, BuildingClass_ActiveClickWith, 0x6)
 
 	if (pThis->GetTechnoType()->UndeploysInto)
 		pThis->SetRallypoint(pCell, false);
-	else if(pThis->IsUnitFactory())
+	else if (pThis->IsUnitFactory())
 		pThis->SetRallypoint(pCell, true);
 
 	return 0x4437AD;
@@ -182,3 +183,16 @@ DEFINE_HOOK(0x44377E, BuildingClass_ActiveClickWith, 0x6)
 // issue #232: Naval=yes overrides WaterBound=no and prevents move orders onto Land cells
 // Author: Uranusian
 DEFINE_LJMP(0x47CA05, 0x47CA33); // CellClass_IsClearToBuild_SkipNaval
+
+// bugfix: DeathWeapon not properly detonates
+// Author: Uranusian
+DEFINE_HOOK(0x70D77F, TechnoClass_FireDeathWeapon_ProjectileFix, 0x8)
+{
+	GET(BulletClass*, pBullet, EBX);
+	GET(CoordStruct*, pCoord, EAX);
+
+	pBullet->SetLocation(*pCoord);
+	pBullet->Explode(true);
+
+	return 0x70D787;
+}